+++ src/components/Gif/styles.css
@@ -21,7 +21,7 @@
 	bottom: 0;
 	left: 0;
 	background-color: rgb(0 0 0 / 75%);
-	color: #fff;
+	color: var(--paragraph);
 	text-decoration: none;
 	border-radius: 0 0 4px 4px;
 	width: 100%;
