+++ src/components/Modal/styles.css
@@ -23,7 +23,7 @@
 	height: 80vh;
 	width: 100%;
 	max-width: 25rem;
-	background-color: #1d1e22;
+	background-color: var(--card-background);
 	border-radius: 0.25rem;
 	padding: 2rem;
 }
