+++ src/components/Spinner/index.js
@@ -1,13 +1,5 @@
 import "./styles.css";
 
-const Spinner = () => {
-	return (
-		<div className="Spinner">
-			<svg className="loading-spinner" viewBox="0 0 50 50">
-				<circle cx="25" cy="25" r="20"></circle>
-			</svg>
-		</div>
-	);
-};
+const Spinner = () => <div className="Spinner"></div>;
 
 export default Spinner;
